+++ src/scenes/parallax/layerData.ts
@@ -1,0 +1,131 @@
+let LayerData = [
+    [
+        {
+            "textureName": "back1.png",
+            "position": {
+                "x": 120,
+                "y": 303
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "back2.png",
+            "position": {
+                "x": 20,
+                "y": 230
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "back3.png",
+            "position": {
+                "x": 210,
+                "y": 200
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "back4.png",
+            "position": {
+                "x": 300,
+                "y": 130
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        }
+    ],
+    [
+        {
+            "textureName": "mid1.png",
+            "position": {
+                "x": 120,
+                "y": 303
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "mid2.png",
+            "position": {
+                "x": 20,
+                "y": 230
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "mid3.png",
+            "position": {
+                "x": 210,
+                "y": 200
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        }
+    ],
+    [
+        {
+            "textureName": "front1.png",
+            "position": {
+                "x": 120,
+                "y": 303
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "front2.png",
+            "position": {
+                "x": 20,
+                "y": 230
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "front3.png",
+            "position": {
+                "x": 210,
+                "y": 200
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        },
+        {
+            "textureName": "front5.png",
+            "position": {
+                "x": 120,
+                "y": 330
+            },
+            "scale": {
+                "x": 0.8,
+                "y": 0.8
+            }
+        }
+    ]
+];
+
+export { LayerData };