+++ src/scenes/preloader.ts
@@ -1,5 +1,5 @@
 import * as Phaser from "./../../node_modules/phaser/dist/phaser.js";
-import { PlayScene } from "./play";
+import { PlayScene } from "./gameplay";
 import { game } from "./../game";
 
 export class LoadScene extends Phaser.Scene {
@@ -16,7 +16,7 @@ export class LoadScene extends Phaser.Scene {
         let splash = this.add.image(game.canvas.width/2, game.canvas.height/2, "atlas", "splash.png");
         splash.setScale(game.canvas.width/splash.width, game.canvas.height/splash.height);
 
-        game.scene.stop("load");
-        game.scene.start("play");
+        game.scene.stop("preloader");
+        game.scene.start("gameplay");
     }
 }